added calculations to desired client size
I added a calculation to give me a window that encases a client size of
known area.  Previously, I had hard coded the window size, but the
result was an unknown client area size that was slightly smaller because
of window borders.

With the new code, the client size is specified, and the window is
adjusted to compensate.